backend/requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
requests==2.32.3
